moved crypto to correct place and moved login logout method to Page class as it is static. Made all suggestions from git
