+++ src/main/resources/Page.properties
@@ -1,3 +1,4 @@
+homepageURL = https://parkme.com/
 testCity = New York
 incorrectUsername = kiruha.testing
 incorrectPassword = 08642
