+++ src/test/java/TestBase.java
@@ -9,11 +9,12 @@ public class TestBase extends Page {
     public HomePage homePage = new HomePage();
     public SearchResultPage searchResultPage = new SearchResultPage();
     public LoginPage loginPage = new LoginPage();
+    public Page page = new Page();
 
     @BeforeTest
     @BeforeMethod
     public void openHomePage() {
-        driver.get("https://parkme.com/");
+        driver.get(Config.getSetting("homepageURL"));
     }
 
     @AfterTest
